feat: I created the full HTML code for the video.
I copied and pasted all the code. I understand most of what it does, but I haven't taken notes because I haven't studied it yet. I don't know if it works or if it's correct. I haven't connected the port or opened the test server. I'll wait to see how it looks on the production page.